adds sidebars definitoin to footer builder, keep the same number of sidebars as in the current neve version
